Moved TopicCredentials.cs out of the implementation directory.
